+++ src/test/kotlin/com/github/vhrabar/issuehub/IssueHubPluginTest.kt
@@ -3,7 +3,6 @@ package com.github.vhrabar.issuehub
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
 
 class IssueHubPluginTest : BasePlatformTestCase() {
-
     fun testBundleMessageResolves() {
         assertEquals("IssueHub", IssueHubBundle["name"])
     }